backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pymongo==4.6.0
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
pydantic>=2.0.0,<2.4.0
pypdfium2==4.0.0
pdf2image==1.17.0
Pillow>=9.4.0,<10.0.0
gunicorn==23.0.0

# Video agent dependencies
google-cloud-aiplatform>=1.36.0
opencv-python>=4.8.0
requests>=2.31.0
pytube>=15.0.0
xmltodict>=0.13.0
google-genai==1.1.0

# Additional dependencies found in codebase
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0
yt-dlp>=2023.0.0
aiohttp>=3.8.0
numpy>=1.24.0
pytesseract>=0.3.10
boto3>=1.34.0 # Added for Cloudflare R2 interaction
redis>=6.0.0 # Added for Redis caching
aioredis>=2.0.0 # Added for async Redis operations
firebase-admin>=6.2.0 # Added for Firebase authentication and Firestore database
twelvelabs>=0.5.0 # Added for Twelve Labs video search and indexing
